+++ typescript/agentkit/src/action-providers/index.ts
@@ -13,3 +13,4 @@ export * from "./twitter";
 export * from "./wallet";
 export * from "./customActionProvider";
 export * from "./alchemy";
+export * from "./moonwell";